refactor: move S3LazyRead to S3IoPolicy
